adapt drumkit property edit for song kits
in the Main Menu > Drumkit the "Save" option was removed (as saving a kit associated to a song is done by saving the song) and "Save As" was renamed to "Save To Sound Library" in order to make more evident that this action is writing the current kit to the Sound Library.

All drumkit property edits for the kit of the current song can now be undone/redone using the `SE_switchDrumkitAction` and do not result in any XML files other than the current songs being touched. The name of the kit an be set to an arbitrary non-empty name.

The drumkit image of a song drumkit is stored in our cache folder on user-level. When saving the kit to the Sound Library or exporting it, this file will be included too.