Revert "DPD: propagate license for song kits"
This reverts commit d31b99302ca19a3e608ca9365a6ea7924ab916cd.

When thinking about it again, propagating the license right away is probably a bad idea. We do not do this for the drumkit name either, since this would compromise sample loading of imported instruments. But then instruments get into a somewhat dirty state. In addition, the drumkit the instrument was imported from could very well be still in development and changes of its license should been taken into account when saving/exporting the songs kit at a latter point in time.

License propagation is still done when storing a copy of the kit to the Sound Library or when exporting it