DPD: always ask before overwriting kit
when duplicating a kit, save the song kit to the Sound Library, or just edit the properties of an existing kit, Hydrogen should always ask for consent before overwriting an existing drumkit. Usually this is not done frequently, so, this question should not be too annoying.

This also more clearly indicates the difference between storing a song kit and a kit in the Sound Library