Drumkit: support absolute image paths
similar like sample paths in instrument(layers) we do now support two types of image paths. a) plain filenames of images that are expected to be found in the drumkit folder and b) absolute paths to an image.

b) is required when storing the image associated with a song kit in the cache folder of Hydrogen at user-level.