Drumkit: fix per component export
in case two components share the same name.

Previously, selecting a component for export was just done by its name. But this fails in case of components sharing the same name. Instead, the component id is now used to specify a component.

In addition, the same unique labels used in the `InstrumentEditor` are now used in the `DrumkitExportDialog` as well