InstrumentEditor: can be now enabled and disabled
Previously, when the selected instrument was a `nullptr` after e.g. a drumkit change, the instrument name label was updated to "NULL instrument...", the selected layer and the wav display were reset. But most of the widgets still remained intact and using them resulted in crashed (for most of them).

Instead, whenever an instrument is `nullptr`, all the widgets of the `InstrumentEditor` are now cleared and disabled. This very neatly shows the intention of this GUI part to be unusable in the current state.

(I am not sure right now if null instruments only happen due to errors or can be reached via our regular UX. But I remember encountering them from time to time. Anyway. Now we could also think about using them on purpose)