NsmClient: do not guard get_instance()
using an assert.

When using rubberband and having OSC support compiled into Hydrogen, `TransportPosition:setBpm` will trigger a resampling of all samples whenever a tempo change is encountered. This can also during the tear down of `Hydrogen` and specifically of the `AudioEngine` happen when using tempo markers. The `AudioEngine::reset` routine might relocate to a column using a different tempo. But by then the `NsmClient` is already deleted - which is a good thing because it is more high-level than the `AudioEngine`. But since `Drumkit::recalculateRubberband` checks for an active NSM session, the `NsmClient` singleton is accessed and the assert is triggered. Just accessing it does not lead to any problems since `Hydrogen::isUnderSessionManagement` is already checking for a nullptr. I thus just remove the assert