NSM: remove drumkit linking
Previously, the kit last loaded in a session was linked to the drumkit folder. This was done to comply with the NSM standard that archiving a session folder - `tar -z` - should include all data necessary to reproduce the session. However, the kit linked might only contain a subset or even zero of the required samples. Whenever a new layer was added or an instrument drag-drop imported, the corresponding sample was located _outside_ of the linked drumkit and would not appear in the archive. The only way to circumvent this problem was to save the current drumkit back to the Sound Library again.

By now, drumkit handling in H2 has matured a bit and there is a more neat way of implementing the behavior above. The concept of "last loaded drumkit" was dropped because it lead to inconsistencies. Instead of linking a kit in the folder the user will be present (done in a future commit) the option to save the current songs kit into the session folder. During startup/ session switching all drumkits found in the session folder will be loaded in the Sound Library as session kits