check audio driver pointer
The audio driver can be changed/switched and in the process it might become `nullptr`. Not on a regular basis tough. But due to another bug during local development it did become null during startup and Hydrogen crashed with almost no way to bring it up again.

To ensure this never happens in productive code, I introduced checks for `nullptr` on all calls to its members. It neither atomic nor guarded by a mutex. But it is already a more resilent and, I think, sufficiently save