SLED: Fix file browser (#1927)
Exporting a drumkit requires the user to select a folder to export it to. (Not the name of the result!)

Since 1.2.0 this was broken at least for Windows in such a way the user was able to select files as well. In addition, double clicking a folder was not selecting but opening it. Also, users were able to see files in the file browser too. In the end, the path had to manually sanitized in our export dialog to ensure it was pointing to an existing folder.

Now, only folders are shown again and they can be selected using the OK buttonn