AppVeyor/Win: hard-code Qt5 version
for some reason the Qt5 package could not be found in the Win32 build https://ci.appveyor.com/project/mauser/hydrogen/builds/48874619/job/3cptmn98gvsgu2sb . It happened several times in a row but it is still quite strange. After all, the package for mingw64 follows the exact same naming conventions and it could be found there.

The package now carries the full version identifier in order to check whether this caused the problem.