AudioEngine: no process while not ready
probably due to some upstream JACK bugs (not sure) Hydrogen does occassionally segfault when stopping the `JackAudioDriver`. Turns out JACK has a number of issues with thread joining. During teardown of the driver we must ensure it does not do any communication, like locking the `AudioEngine` or enqueuing a message in our `Logger`. Otherwise the thread calling `jack_client_close` might segfault.

Now, the callback of the `AudioEngine` checks its state twice. Once without locking (in order to `return` when it is not ready yet/anymore) and another time while locked to properly check the state. It is a little dirty. But I do not see what else we can do from within the Hydrogen code base