AE: add drivers to log messages
since I tweaked the issue template and it is asking the user to attach a log as well as provided a description in the Wiki, we have a lot more error logs in new issues. Which is definitely good.

Still, they could be more expressive.

With this patch the audio and MIDI drivers used are added to all log messages of the AudioEngine. This might help to interpret the logs and spot errors more easily