AE: tweak debug log levels
there are still quite a number of commented log messages within the `AudioEngine` class. They are key for debugging stuff changed for the 1.2.0 version and save a lot of time. However, things seem to be more and more stable and I had not used them in months.

Previously, they had all sorts of log levels including debug, info, and warning in order to better read the resulting log due to the different colors associated to each level. Those all have been set to debug now as they are definitely debug level and in order to not confuse new developers.