Legacy: tweak load_playlist
instead of creating a fresh `Playlist` and pass it as argument to `Legacy::load_playlist` while, at the same time, returning this pointer, a new playlist is created in the legacy function itself. It is much clearer.

The `Playlist` constructor has to be made public but I intend to refactor the code for the playlist to not be a singleton anymore anyway.