Playlist: start to move logic into core
most of the logic handling `Playlist` is found in `PlaylistEditor`. But this is a) not a good place with respect to separation of concerns and b) prevents headless Hydrogen version, like `h2cli` and `h2player`, make proper use of the current playlist.

This commit lays some foundations to move these logic into `Playlist` and make it accessible to the GUI/OSC via `CoreActionController`