CAC: unify signatures + asserts
all public member functions of the `CoreActionController` do now have `bool` as return type indicating success or failure.

In addition, an assert and sanity check was introduced to each function checking whether the `Hydrogen` instance is present. This is important as the members are now static and they must not be used in case the core is not ready yet (during startup or shutdown).