Playlist: move song loading into core
previously loading a song from the current Playlist was done by the GUI with the core just triggering a dedicated event upon receiving a MIDI or OSC message. This is not nice because H2 should work headless at some point. (`h2cli` did contain a workaround by reimplementing the GUI functions. But those already got out out of sync.) Instead, loading a song from a Playlist is now done entirely within the core and the existing event was repurposed to just show a dedicated status message in the GUI.

In addition, scripts associated to a song in the Playlist are now always executed after the song got loaded