Playlist: tweak song exists check
`PlaylistEditor` does now check again whether a song in the playlist exists and highlights the corresponding grid cell if not.

`PlaylistEntry` was refactored into a class and does now use getters and setters. This way we can ensure its song path is not altered without rechecking for the underlying files existence.