add CLEAR_SELECTED_INSTRUMENT
and drop the MIDI endpoint `CLEAR_INSTRUMENT` (the OSC counterparts, however, is still available).

This was done since it is infeasible to assign the clearing of instruments to e.g. a knob on a MIDI devices. Clearing both instrument 0 and 10 would result in all instruments inbetween to be cleared as well