PlaylistEditor: add separate undo stack
our main undo stack in `HydrogenApp::m_pUndoStack` is flushed whenever a new song is loaded. This is itself a great idea but just not all too handy when it comes to undoing stuff done in the `PlaylistEditor`.

Instead, `PlaylistEditor` has now a separate undo stack, an undo node in its menu, as well as its own undo history view