CAC: write empty string on new file
in the slots of the `Preferences` for restoring the previous song and playlist we now store an empty string in case a new song/playlist was opened by the user. If there are unsaved changes, they will be stored as an autosave file for the empty song/playlist.

this makes recovery of empty files much easier. And after all, it was an explicit user interaction with the intent to open a new file. It should not surprise that it will be opened during next startup.