use sf_wchar_open on Windows + libsnd logging
over `sf_open` for opening a sndfile file. The former takes an UTF16-BE encoded string. This way we can guarantee all characters provided in via the UI - and stored as UTF16 in Qt internally - end up properly in the filename.

This fixes both loading samples from files containing non-ASCII characters in its absolute path in case a mismatching OS encoding was set and allows to export songs to arbitrary filesnames/paths