SL: notify on import/export failure
Currently `libarchive` is not able to handle UTF-8 filepaths which could not be encoded in the users current encoding on Windows well.

That it not ideal. But I also do not see a nice solution. (We could make Qt instead of libarchive open the file and just pass the file pointer to libarchive. But this does not feel like a good and sustainable solution). Instead, we now have the official limitation that drumkit archives must not carry non-ASCII characters in order for the kit to be supported on all platforms.

In case importing or exporting failed due to encoding issues, a dedicated dialog will the user what went wrong

In addition, import failures on the command line do now result in a non-zero return code.