Remove autosave files when user declines to restore from them
When Hydrogen starts up, if a file has been autosaved the user will
be offered the opportunity to restore from that file. If they decline
it, Hydrogen should delete the autosaved file; it may be soon overwritten anyway by new autosave data.

This avoids the annoying cycle I often find myself in:
   1. edit file
   2. quit Hydrogen discarding edits (but autosave remains)
   3. open Hydrogen
   4. hit "Don't Save" because I didn't want those edits
   5. listen to original file without modifying (hence no autosave, but old autosave remains)
   6. quit Hydrogen (silently)
   7. goto 3

Also, when opening a file and restoring from its autosave, Hydrogen
should count this as the song being Modified because the current
content is from the autosave and is not the same as the content of
the named file.