add -T/--log-timestamps CLI option
to add timestamps to all log messages. This is required to correlate log messages from two different instances of Hydrogen within the JACK timebase integration test.

In addition, the CLASS::FUNC prefix was put in squared brackets in order to more clearly distinguish them from the actual logging message. Especially for callbacks this was not that easy