OscServer: expose server thread
this way the program using libhydrogen-core, e.g. an integration test or `h2cli`, can add new, custom handlers. This allows for less invasive testing of core functions without the need to introduce further and further OSC endpoints just for unit testing.

It is not a really beautiful solution yet as `OscServer::generic_handler` is already install when creating the OscServer instance. This way we get a false positive error message for all OSC handlers installed later on. But as this is only intended for "internal" development and testing I think it is fine