Drumkit: fix UTF-8 import
after inspecting the source code of `libarchive`, it turns out the library _does_ contain dedicated functions for the UTF16 Windows API. Similar to `libsndfile` they are called `*_w`. They just do not appear in the man pages as those seem to be built on BSD and the Windows part of the API was removed by ifdefs.

Import of an ASCII drumkit located in a folder containing Cyrillic characters does already work.

Exporting a kit already produces a .h2drumkit file and does not report an error, but the Cyrillic characters in the kit name are lost. This needs to be refined