h2cli: rework CLI parsing
the entire CLI parsing in `h2cli` was binned and replaced with a variant of the code used in `hydrogen`. This make both the code base more easy to maintain as well as aligns the UX. In particular the `-V` option in `h2cli` was not able to handle whitespaces between flag and argument. Small details, but there is no need for things to diverge that much.

In addition, the long option `--drumkit` in `h2cli` was renamed to `--kit` as used in `hydrogen` since release 1.0.0-beta1