DiskWriterDriver: indicate completeness
The `DiskWriterDriver` is instantiated to write an audio file to disk. Whether it is done or not was till now indicated using `EVENT_PROGRESS` in the `EventQueue`. But this had several downsides. Our events are designed to be one-way messages between core and GUI. There is no way the core knows whether export was completed or failed. To complicate things, events are poped from the `EventQueue`. So, reading them at several places is off the table without rewriting some fundamental parts of Hydrogen. Instead, a bool member of the `DiskWriterDriver` class is now used to indicated success/failure.

Within our unit tests `TestHelpers::exportSong` used `EVENT_PROGRESS` to determine whether the pthread of the `DiskWriterDriver` was done. But this is bad design as the program wrapping `libhydrogen-core` must not use events in any other place. Instead, this helper function does now rely on the bool member introduced in this commit and the integration test binary is able to interpret and act on all EVENTs