AudioEngine: fix JACK tick rounding
When converting between ticks and frames a tiny mismatch (rounding error) is encountered. Internally this one is always taken into account in order to allow for a seamless conversion between the two without glitches. However, when using the `JackAudioDriver` relocation is done via the JACK server and the mismatch is lost. Relocating to tick 1.9999 -> converting it into frames -> sending this frame to the JACK server -> using the frame received from JACK server -> converting it back into a tick -> 2.0. We encounter a glitch. In order to avoid this, the tick is floored.

Remembering the mismatch from the initial tick -> frame conversion is of no use either. Any other JACK client can initiate a relocation too and we have to way to be sure the received value is the one sent by us.

Previously, flooring was done on the value received from the JACK server. This fixed the glitch but resulted in a loop of log messages as the frame assigned internally differed from the one sent by the JACK server (due to the flooring), which in turn was considered a second relocation. Now, flooring is done on the initial value sent to the JACK server.