jack: rework Timebase state handling
Unfortunately, the JACK server sends spurious BBT information when stopping transport: it switches to no BBT support till playback is running again. This means we loose access to the masters bpm and have a spurious GUI change (button state) between stopping and playing.

To avoid this, we buffer the last tempo and timebase state and skip treating a stopped JACK transport pos on frame 0 with respect to Timebase information.