AppVeyor: drop JACK2 support for Windows32
I do still see a `mingw-w64-i686-jack2` package in https://repo.msys2.org/mingw/mingw32/. But both locally and in the AppVeyor pipeline the package can not be installed and querring  does yield no results. We have no choice but to disable JACK support for the Windows 32bit build for now.