jack: rework Timebase state handling
`JackAudioDriver::m_nTimebaseTracking` was converted into an enum class
`JackAudioDriver::m_timebaseTracking` with states `Valid` and `OnHold`. Hydrogen
is now able to handled relocation of non-master JACK clients while not being
master itself well.

While Hydrogen can unregister as timebase master on its own, it can
not be observed directly whether another application has taken over
as timebase master. When the JACK server is releasing Hydrogen in the
later case, it will not advertise this fact but simply will not call the
JackTimebaseCallback() anymore. But since this will be called in
every cycle after updateTransportPosition(), we make the former set
the tracking state to #Valid and the latter to #OnHold. If we
encounter a second process cycle with #OnHold, we have been released.

A second use case is relocation triggered by a JACK client other than
the Timebase master. The JACK server will not have any corresponding BBT
information at hand and distribute the frame information without the
BBT capability. The process cycle afterwards the master starts again
to send BBT. This intermediate state will be covered by #OnHold too,
as we do not want the timebase state (and timeline support) to glitch
on each relocation. Instead we cache the last tempo and pretend to
still have BBT information. will be updated accordingly.