jack: fix pattern selection in timebase callback
when making pattern sizes more diverse and adding a virtual pattern to the JACK timebase integration test song it turned out the timebase callback of `JackAudioDriver` did not select the right pattern. Instead of using the playing ones just the one presented to the user was used. In addition, the `AudioEngine` was not locked while putting BBT information together. This is starting to be embarassing.

Now BBT information is derived based on the longest of the currently playing patterns (including virtual ones). In case there are several of the same (max) size, the first one encountered is used.