Hydrogen: lock AudioEngine while setting song
for some reason the audio engine was not properly locked when setting a different song. At least not all the way to ensure everything is consistent again. For just a split millisecond the process callback of the AudioEngine could even be called with the current song set to nullptr.

I guess there was no intention behind this and it is just a bug I introduced while rewriting handling of `AudioEngine`.