jack: drop JackBBTSyncMethod and compareAdjacentBBT
Previously we had two different approaches to react on BBT (bar, beat, tick) information provided by the JACK server: 1) const measure and 2) identical bars.

In 1) we convert BBT into a tick position and relocate to this very point. If patterns in Hydrogen have a different length than the measure used by the Timebase master, bar: 2 will not necessarily result in Hydrogen relocating into pattern 2. But on the upside, transport is consistent. There are no sudden jumps and it does not matter whether the Timebase master relocates to BBT X or lets transport roll from an arbitrary prior point till that position. Hydrogen will be in the exact same position in both cases.

In 2) the bar information would be mapped to our columns. This means a mismatch between pattern length and measure in the master would either result to jumps during playback or inconsistent position when either rolling or relocating to position X. That is no good UX. That is why we suggested in the manual to ensure pattern lengths always match measures in the master. But when doing so this approach yields the same results as 1). So, I just drop it to make things more simple. `JackAudioDriver::compareAdjacentBBT` was responsible for check for changes in the BBT . This is redundant now as BBT information is expected to always be consistent with the frame provided by JACK and we already check this one for changes. Luckily the function did not do anything in the last releases due to a bug as it would have (further) messed up transport