jack: fix relocation beyond the end of song
It is important to relocate to the beginning of the song. If we would stay at the end or beyond, Hydrogen would stop every attempt to start playback again. And it is most probably not obvious to the user why it does so.

Also, `JackAudioDriver` needs to trigger an `EVENT_RELOCATION` itself as it uses the internal `AudioEngine::locate` directly