AE: fix updateTransportPosition
`AudioEngine::updateTransportPosition` was meant to be called for an arbitrary `TransportPosition` without any side effects for the `AudioEngine` or other parts of Hydrogen. All intended side-effects are applied by previously checking whether the supplied `TransportPosition` is `AudioEngine::m_pTransportPosition`.

But of these checks a number was missing. This caused spurious tempo change and other events triggered by `AudioEngine::m_pQueuingPosition` as well as an update of all enqueued notes when the latter encountered a tempo changed. This is itself most probably nothing harmful but duplicated computations.

;