jack: use provided frame in TimebaseCallback
there had been a design flaw in `JackAudioDriver::JackTimebaseCallback`. Instead of just using the BBT information of the current transport position in the AudioEngine, the BBT information has to be provided for frame the JACK server supplied to the callback.

It works a bit but not flawlessly yet. Whenever a XRun occurs, the frame information provided as function input IS CHANGED BY THE JACK SERVER. So, while calculating the BBT information for a frame the JACK server just replaces that very frame. WTF? We have to take this into account because it is bad and it makes our integration test fails as soon as we activate a bunch of log messages