tests: rework BBT relocation testing
Previously we kept track of the number of relocation attempts in order to check whether are relocation loops - Hydrogen regards the same position a relocation over and over again. But this did not work.

For one when triggering relocation in the listener, no valid BBT position will be provided and relocation will work using the classic frame approach. It can only be tested in the additional BBT relocation test where it is now done explicitly