jack: rework BBT relocation
The first time we get a valid BBT information, we consider this a relocation and calculate the new position based on BBT.

I have to admit I had this already on my todo list once or twice but abstained from it since it has a really bad UX consequence for Hydrogen: clicking on a position on the ruler is not guaranteed anymore to make Hydrogen relocate to this position.

The reason is as follows: Hydrogen calculates the frame for the desired position based on the current tempo and measure information provided by the JACK Timebase master - which will lead to wrong results in case of tempo and measure changes. But that is all we can base or work on. The Hydrogen sends this frame to the JACK server, which will broadcast it without any BBT info. We relocate to the desired location. But in the next processing cycle the Timebase master fills the BBT information corresponding to the frame and broadcasts it.

If we a) consider this a relocation and calculate the new position based on BBT we loose the ability to relocate transport from within Hydrogen. If we b), just use the tempo information and remain at the same location (as done previously) we only use part of the BBT information and are in a wrong and inconsistent transport state. To c) do not act on the BBT info at all, we already support by disabling BBT support within the Preferences.

As far as I see it we can only go with a). But it is ugly and I am a little bit annoyed of the bad design of the JACK server and how much work it forces me to do when still the results do behave soo badly. But apart from a disclaimer in the manual there is not much we can do here. (If there would be a way to request frame information for a particular BBT point, everything would be fine.)