AudioFileBrowser: fix layout scaling
scroll bars have been introduced using which the dialog is still usable below its minimum size. In addition, a maximization button was added to allow forcing the window below its minimum size when using a larger scaling factor.

The `QTreeView` defined in the _UI.ui file is now used directly instead of creating yet a second as its child. This allows to harness full screen size in the tree view when expanding the widget.