AboutDialog: fix layout scaling
scroll bars have been introduced using which the dialog is still usable below its minimum size. In addition, a maximization button was added to allow forcing the window below its minimum size when using a larger scaling factor.

The layout of the tab widget was changed as well to harness all screen size for the displayed text when enlarging the dialog.