rework DrumkitMap and instrument type handling
as outlined in "Drumkit-indepedent patterns proposal V2" (#1984) mappings between instruments and instrument types are now 1 : 1. As a result we can just integrate a type property to each instrument and drop the `m_pDrumkitMap` and its fallback in `Drumkit`. The `.h2map` files are now only shipped by us an used in case no type was set for the kit. This eases type handling a lot