SLDatabase: rework type sorting
It was a nice idea to sort all instrument types by the number of their occurrences in the combo boxes of Drumkit > Properties. But this very quickly becomes unfeasible and one does not find anything anymore. As we do not was to make things overly complex - e.g. filtering or different ways to sort content - we will just stick with alphabetical ordering