tests: add assertion to check XML file equality
with the existing `H2TEST_ASSERT_FILES_EQUAL` macro and corresponding function it is impossible to properly compare our XML files on byte level. The XML standard seems not to define a concept of order for XML element attributes and the namespace ones of the root element are written in a seemingly random fashion. This results in spurious false positives.

The new `H2TEST_ASSERT_XML_FILES_EQUAL` reads in the two XML files, serialize the parsed document, and perform a string comparison. Attribute order is no longer an issue.

The comparison works on the whole content but assertion message is providing the exact line something goes wrong:

1) test: XmlTest::testDrumkitMap (F) line: 445 /home/phil/git/hydrogen/src/tests/XmlTest.cpp
XML files differ at line [10]:
	expected:   <instrumentID>1</instrumentID>
	actual  :   <instrumentID>2</instrumentID>
- Expected: /home/phil/git/hydrogen/src/tests/data//drumkit_map/ref-duplicates.h2map
- Actual  : /tmp/hydrogen/ref-saved.h2map