tests: use cmake to determine source and binary folder
over a custom solution using a system call to `git` (since we are using `cmake` anyway).

The variables were added to `src/core/config.h.in` though they are not directly used within the core part. But I think it is better to have a single source of truth for all cmake-based variables which can be imported in all other parts