tests: assert timeout in CLI tests
damn. It turns out the unit tests did not take significantly longer due to the CLI tests. There was just a bug in the CLI causing the call to run till timeout. Well, but this is also what we are writing tests for.

I added asserts checking whether an error occurred in process execution or whether the process was timed out