tests: add consistency tests for .h2pattern files
previously, it was just tested whether or not a pattern could be saved or loaded. Now, we check whether saving a pattern gives the exact same XML file (breaking the current format should be an explicit thing and the intention should be shown by fixing the corresponding test). Also, I added a unit test for loading legacy formats