tests: add XML consistency test for playlist
previously, it was just tested whether or not a playlist could be saved or loaded. Now, we check whether saving a pattern gives the exact same XML file (breaking the current format should be an explicit thing and the intention should be shown by fixing the corresponding test).

The test, however, fails! This is because loading of relative paths in a playlist is done relative to the pattern file itself but writing is done relative to the user playlist folder