tests: add XML consistency test for h2song
previously, it was just tested whether or not a song could be saved or loaded. Now, we check whether saving a pattern gives the exact same XML file (breaking the current format should be an explicit thing and the intention should be shown by fixing the corresponding test).

The test, however, are partially commented for now as they need to be portable first.