Split Pattern/Note loading and instrument assignment
Loading a `Pattern`, `PatternList`, or `Note` from XML and assigning an `Instrument` to (all contained) note(s) are now separated steps. Both `Pattern` and `PatternList` do now have a method called `mapTo` which takes a drumkit and takes care of assigning an instrument to the individual notes. This will be later used to implement the drumkit switching as well