AppVeyor: no ANGLE on Windows 32bit
as for JACK2 there is also no `angleproject` package for 32bit Windows within the `pacman` repos. At least as far as queries from the CLI both locally and within AppVeyor are concerned. In the web interface the package seems to be still available.

This starts to smell like bit rot. Lets see how long we can provide 32bit support