data/drumkit_maps: tweak toms and crashes
while doing the first iteration of the instrument types I used the names of the instruments to derive the type. But especially with the toms this lead to problems. Some kits seem to use the terms "high" and "low" to describe the pitch of the sound relative to each other instead of the instrument itself. This lead to many floor toms being labeled as low tom.

I am not an expert nor a drummer but having "Tom Mid", "Tom Low", and "Tom Floor" for default kits and reserving "Tom High" for kits with additional higher pitched seems to be a reasonable thing