Playlist: drop name element
our `.h2plyalist` file had a redundant and annoying element called `name`. It was always set to the name of `.h2playlist` file. The exact same name. And it was not even used to fill `Playlist::m_sFilename`.

As it breaks consistency tests when using random tmp file names, I removed it. Starting from now we do not write this element anymore and it is marked optional in the corresponding XSD file. (This itself would invalidate new `.h2playlist` files for older Hydrogen versions. But since this already happening due to `formatVersion`, we do not loose anything in here)