Playlist: fix relative song path saving
turns out saving song files relative to a playlist was not working (most of the time). While import resolved the song path relative to the location of the actual `.h2playlist` file, saving wrote them relative to the users playlist folder. This means, only in case the `.h2playlist` file was located in this very folder loading did work.

Song files are now always resolved relative to the actual `.h2playlist` file.