Refactor Preferences
Instead of having just the `Preferences` singleton to load system-level or user-level settings into, we now adhere to the pattern we already use for e.g. `Song` and `Playlist` and provide a `load()` method to retrieve a `Preferences` instance without altering anything within Hydrogen, a `save()` method and `saveCopyAs()` to store changes, as well as `load*`, `set*`, `save*` routines in `CoreActionController`, which can be used from e.g. the GUI part.

Note that there are still side effects in `Preferences::load` since `MidiMap` is (still) a singleton.

Using the pattern above will enable use to both provide unit tests for preferences XML file, like format consistency, and will allow use to redo/undo changes done in the `PreferencesDialog`.

At some previous point the whole concept of drumkit loading seems to have been as follows: First, some default values were set in the constructors, then the system-level settings are loaded and assigned, and last of all the user-level ones. But since there are no side effects and both loading on system and user level yields fully fledged `Preferences` instances, loading the system one is redundant in case loading on user-level worked. System-level will now only be used as a fallback.

In addition, upgrading the user-level preferences file on startup was dropped as well. It is also kinda redundant since loading it produces an updated version in memory and when quitting Hydrogen this updated version is written to disk anyway.

I also refactored large parts of `Preferences::load`. Now, it uses the values assigned in the constructor(s) as default ones. This was already done in a lot of places. But not all. Occassionally the default values did already diverge.