Preferences: align ordering of members
for initialization in constructor and as definitions in header. This is done to make sure none of them is missed (essential for the definition of a copy constructor, which comes next).

Also, a couple of trivial initalization have been missed.

In addition, the `hydrogen.conf` file does now contain one additional field called `playSamplesOnClicking`. This variable was missed and never written to disk.