Preferences: drop m_bUseTimelineBpm
this was kept as some sort of fallback when loading older song. In more recent ones, timeline activation and usage is both stored in and read from .h2song files. The definition in `Preferences` was redundant and caused loading a free standing `Song` to have side effects by leaking its timeline activation into the current instance of `Preferences`