rename `version` -> `userVersion`
in `.h2song`, `.h2pattern`, and `drumkit.xml` files.

There is already a `version` element in `.h2song` (and `hydrogen.conf`) indicating the semantic version the file was created with. That is a problem since the former value would not even be compatible with the integer type of `userVersion`. In addition, we do not want to make backward and forward compatibility more complex than it already is.

As a result the newly introduced `version` will be renamed to `userVersion`. But the corresponding member functions are still called only "version" (as the existing `version` is not read from file and just a transient property)