DPE: more resilient not drawing
in case a note within a pattern did not had an instrument (broken all-instruments-copy and all-instruments-paste in this case) caused a segfault since `DrumPatternEditor::drawPattern` tries to access the instruments of all notes.

This will be overhauled at a later point. For now it is just important we do not get any segfaults on `master`