PatternEditors: decrease cursor size on res. `off`
the default cursor size is twice the distance between grid line in 1/64 resolution. This is nice, because it allows both for precise keyboard-selection and is still big enough to properly see it.

But when setting resolution to `off` things change. Resolution is now 1/192 and the cursor is just too big. It is impossible to select individual notes in densly populated patterns using keyboard.

Instead, now the size of the cursor is decreased in case resolution is set to `off` to cover twice the distance between grid lines. This yields the same UX as for 1/64 and allows to properly select individual notes.