tests/XmlTest: improve test of shipped drumkits
previously the shipped kit files were only validated against our XSD definition. But this was useless. Those drumkits will be updated on startup of Hydrogen - in the beginning on the test.

Instead, we now check whether an update took place (check for backup files) and we also load, save, and compare all shipped `drumkit.xml` files to verify we indeed ship the latest defition. Else, we might miss new but optional fields.